refactor: extract devto and qiita workflows
